created critic miner for rotten tomatoes to get prefs for a recommender
